Guitar Pro implementation split up, support also for ties, grace notes, slides and other minor articulations.
