+++ mscore/importgtp.cpp
@@ -47,6 +47,8 @@
 #include "libmscore/volta.h"
 #include "libmscore/fret.h"
 #include "libmscore/instrtemplate.h"
+#include "libmscore/glissando.h"
+#include "libmscore/chordline.h"
 #include "preferences.h"
 
 namespace Ms {
@@ -372,6 +374,66 @@ void GuitarPro::readLyrics()
       for (int i = 0; i < 5; ++i) {
             readInt();
             readWordPascalString();
+            }
+      }
+
+//---------------------------------------------------------
+//   createSlide
+//---------------------------------------------------------
+
+void GuitarPro::createSlide(int slide, ChordRest* cr, int staffIdx)
+      {
+      // shift slide
+      if (slide == 1) {
+            Glissando* s = new Glissando(score);
+            s->setText("");
+            s->setGlissandoType(Glissando::Type::STRAIGHT);
+            cr->add(s);
+            }
+      // legato slide
+      else if (slide == 2) {
+            Segment* prevSeg = cr->segment()->prev1(Segment::Type::ChordRest);
+                  Element* e = prevSeg->element(staffIdx);
+                  if (e) {
+                        if (e->type() == Element::Type::CHORD) {
+                              ChordRest* prevCr = static_cast<ChordRest*>(e);
+                              createSlur(true, staffIdx, prevCr);
+                              }
+                        }
+            createSlur(false, staffIdx, cr);
+            // we should also create a slur between the two notes - to be done when slur PR is merged.
+            Glissando* s = new Glissando(score);
+            s->setText("");
+            s->setGlissandoType(Glissando::Type::STRAIGHT);
+            cr->add(s);
+            }
+      // slide out downwards
+      else if (slide == 4) {
+            ChordLine* cl = new ChordLine(score);
+            cl->setChordLineType(ChordLineType::FALL);
+            cl->setStraight(true);
+            cr->add(cl);
+            }
+      // slide out upwards
+      else if (slide == 8) {
+            ChordLine* cl = new ChordLine(score);
+            cl->setChordLineType(ChordLineType::DOIT);
+            cl->setStraight(true);
+            cr->add(cl);
+            }
+      // slide in from below
+      else if (slide == 16) {
+            ChordLine* cl = new ChordLine(score);
+            cl->setChordLineType(ChordLineType::PLOP);
+            cl->setStraight(true);
+            cr->add(cl);
+            }
+      // slide in from above
+      else if (slide == 32) {
+            ChordLine* cl = new ChordLine(score);
+            cl->setChordLineType(ChordLineType::SCOOP);
+            cl->setStraight(true);
+            cr->add(cl);
             }
       }
 
@@ -1778,997 +1840,6 @@ int GuitarPro3::readBeatEffects(int track, Segment* segment)
 
 
 //---------------------------------------------------------
-//   readMixChange
-//---------------------------------------------------------
-
-void GuitarPro4::readMixChange(Measure* measure)
-      {
-      /*char patch   =*/ readChar();
-      char volume  = readChar();
-      char pan     = readChar();
-      char chorus  = readChar();
-      char reverb  = readChar();
-      char phase   = readChar();
-      char tremolo = readChar();
-      int tempo    = readInt();
-
-      if (volume >= 0)
-            readChar();
-      if (pan >= 0)
-            readChar();
-      if (chorus >= 0)
-            readChar();
-      if (reverb >= 0)
-            readChar();
-      if (phase >= 0)
-            readChar();
-      if (tremolo >= 0)
-            readChar();
-      if (tempo >= 0) {
-            if (tempo != previousTempo) {
-                  previousTempo = tempo;
-                  setTempo(tempo, measure);
-                  }
-            readChar();
-            }
-
-      readChar();       // bitmask: what should be applied to all tracks
-      }
-
-//---------------------------------------------------------
-//   readBeatEffects
-//---------------------------------------------------------
-
-int GuitarPro4::readBeatEffects(int track, Segment* segment)
-      {
-      int effects = 0;
-      uchar fxBits1 = readUChar();
-      uchar fxBits2 = readUChar();
-      if (fxBits1 & 0x10) {
-            effects = 4; // fade in
-            }
-      if (fxBits1 & 0x20) {
-            effects = readUChar();      // effect 1-tapping, 2-slapping, 3-popping
-            }
-      if (fxBits2 & 0x04)
-            readTremoloBar(track,segment);
-      if (fxBits1 & 0x40) {
-            int strokeup = readUChar();            // up stroke length
-            int strokedown = readUChar();            // down stroke length
-
-            Arpeggio* a = new Arpeggio(score);
-            if( strokeup > 0 ) {
-                  a->setArpeggioType(ArpeggioType::UP);
-                  }
-            else if( strokedown > 0 ) {
-                  a->setArpeggioType(ArpeggioType::DOWN);
-                  }
-            else {
-                  delete a;
-                  a = 0;
-                  }
-
-            if(a) {
-                  ChordRest* cr = new Chord(score);
-                  cr->setTrack(track);
-                  cr->add(a);
-                  segment->add(cr);
-                  }
-            }
-      if (fxBits2 & 0x02)
-            readUChar();            // stroke pick direction
-      if (fxBits1 & 0x01) {         // GP3 column-wide vibrato
-            }
-      if (fxBits1 & 0x2) {          // GP3 column-wide wide vibrato (="tremolo" in GP3)
-            }
-      return effects;
-      }
-
-//---------------------------------------------------------
-//   readNote
-//---------------------------------------------------------
-
-bool GuitarPro4::readNote(int string, Note* note)
-      {
-      uchar noteBits = readUChar();
-
-      //
-      // noteBits:
-      //    7 - Right hand or left hand fingering;
-      //    6 - Accentuated note
-      //    5 - Note type (rest, empty note, normal note);
-      //    4 - note dynamic;
-      //    3 - Presence of effects linked to the note;
-      //    2 - Ghost note;
-      //    1 - Dotted note;  ?
-      //    0 - Time-independent duration
-
-      if (noteBits & 0x04) {
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-            note->setGhost(true);
-            }
-
-      bool tieNote = false;
-      uchar variant = 1;
-      if (noteBits & 0x20) {
-            variant = readUChar();
-            if (variant == 1) {     // normal note
-                  }
-            else if (variant == 2) {
-                  tieNote = true;
-                  }
-            else if (variant == 3) {                 // dead notes = ghost note
-                  note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-                  note->setGhost(true);
-                  }
-            else
-                  qDebug("unknown note variant: %d", variant);
-            }
-
-      if (noteBits & 0x1) {               // note != beat
-            int a = readUChar();          // length
-            int b = readUChar();          // t
-            qDebug("          Time independend note len, len %d t %d", a, b);
-            }
-      if (noteBits & 0x2) {               // note is dotted
-            }
-
-      // set dynamic information on note if different from previous note
-      if (noteBits & 0x10) {
-            int d = readChar();
-            if (previousDynamic != d) {
-                  previousDynamic = d;
-                  addDynamic(note, d);
-                  }
-            }
-
-      int fretNumber = -1;
-      if (noteBits & 0x20)
-            fretNumber = readUChar();
-
-      // check if a note is supposed to be accented, and give it the sforzato type
-      if (noteBits & 0x40) {
-            Articulation* art = new Articulation(note->score());
-            art->setArticulationType(ArticulationType::Sforzatoaccent);
-            if (!note->score()->addArticulation(note, art))
-                  delete art;
-            }
-
-      if (noteBits & 0x80) {              // fingering
-            int a = readUChar();
-            int b = readUChar();
-            qDebug("Fingering=========%d %d", a, b);
-            }
-      bool slur = false;
-      if (noteBits & 0x8) {
-            uchar modMask1 = readUChar();
-            uchar modMask2 = readUChar();
-            if (modMask1 & 0x1)
-                  readBend(note);
-            if (modMask1 & 0x2) {         // hammer on / pull off
-                  slur = true;
-                  }
-            if (modMask1 & 0x8) {         // let ring
-                  }
-            if (modMask1 & 0x10) {
-                  int fret = readUChar();            // grace fret
-                  int dynamic = readUChar();            // grace dynamic
-                  int transition = readUChar();            // grace transition
-                  int duration = readUChar();            // grace duration
-
-                  int grace_len = MScore::division/8;
-                  if (duration == 1)
-                        grace_len = MScore::division/8; //32th
-                  else if (duration == 2)
-                        grace_len = MScore::division/6; //24th
-                  else if (duration == 3)
-                        grace_len = MScore::division/4; //16th
-
-                  Note* gn = new Note(score);
-
-                  if (fret == 255) {
-                        gn->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-                        gn->setGhost(true);
-                        }
-                  gn->setFret((fret != 255)?fret:0);
-                  gn->setString(string);
-                  int grace_pitch = note->staff()->part()->instr()->stringData()->getPitch(string, fret);
-                  gn->setPitch(grace_pitch);
-                  gn->setTpcFromPitch();
-
-                  Chord* gc = new Chord(score);
-                  gc->setTrack(note->chord()->track());
-                  gc->add(gn);
-                  gc->setParent(note->chord());
-                  note->chord()->add(gc);
-
-                  // TODO: Add dynamic. Dynamic now can be added only to a segment, not directly to a grace note
-                  addDynamic(gn, dynamic);
-
-                  TDuration d;
-                  d.setVal(grace_len);
-                  if(grace_len == MScore::division/6)
-                        d.setDots(1);
-                  gc->setDurationType(d);
-                  gc->setDuration(d.fraction());
-                  gc->setNoteType(NoteType::ACCIACCATURA);
-                  gc->setMag(note->chord()->staff()->mag() * score->styleD(StyleIdx::graceNoteMag));
-
-                  if (transition == 0) {
-                        // no transition
-                        }
-                  else if(transition == 1){
-                        //TODO: Add a 'slide' guitar effect when implemented
-                        }
-                  else if (transition == 2 && fretNumber>=0 && fretNumber<=255 && fretNumber!=gn->fret()) {
-                        QList<PitchValue> points;
-                        points.append(PitchValue(0,0, false));
-                        points.append(PitchValue(60,(fretNumber-gn->fret())*100, false));
-
-                        Bend* b = new Bend(note->score());
-                        b->setPoints(points);
-                        b->setTrack(gn->track());
-                        gn->add(b);
-                        }
-                   else if (transition == 3) {
-                         // TODO:
-                         //     major: replace with a 'hammer-on' guitar effect when implemented
-                         //     minor: make slurs for parts
-
-                         ChordRest* cr1 = static_cast<Chord*>(gc);
-                         ChordRest* cr2 = static_cast<Chord*>(note->chord());
-
-                         Slur* slur = new Slur(score);
-                         slur->setStartChord(static_cast<Chord*>(cr1));
-                         slur->setEndChord(static_cast<Chord*>(cr2));
-                         slur->setTick(cr1->tick());
-                         slur->setTick2(cr2->tick());
-                         slur->setTrack(cr1->track());
-                         slur->setTrack2(cr2->track());
-                         score->addElement(slur);
-                         }
-                  }
-            if (modMask2 & 0x1) {   // staccato
-                  }
-            if (modMask2 & 0x2) {   // palm mute - mute the whole column
-                  }
-            if (modMask2 & 0x4) {    // tremolo picking length
-                  int tremoloDivision = readUChar();
-                  Chord* chord = note->chord();
-                  Tremolo* t = new Tremolo(chord->score());
-                  if (tremoloDivision == 1) {
-                        t->setTremoloType(TremoloType::R8);
-                        chord->add(t);
-                        }
-                  else if (tremoloDivision == 2) {
-                        t->setTremoloType(TremoloType::R16);
-                        chord->add(t);
-                        }
-                  else if (tremoloDivision == 3) {
-                        t->setTremoloType(TremoloType::R32);
-                        chord->add(t);
-                        }
-                  else
-                        qDebug("Unknown tremolo value");
-                  }
-            if (modMask2 & 0x8)
-                  readUChar();      // slide kind
-            if (modMask2 & 0x10)
-                  readUChar();      // harmonic kind
-            if (modMask2 & 0x20) {
-                  readUChar();      // trill fret
-                  readUChar();      // trill length
-
-                  // add the trill articulation to the note
-                  Articulation* art = new Articulation(note->score());
-                  art->setArticulationType(ArticulationType::Trill);
-                  if (!note->score()->addArticulation(note, art))
-                        delete art;
-
-                  }
-            }
-      if (fretNumber == -1) {
-            qDebug("Note: no fret number, tie %d", tieNote);
-            }
-      Staff* staff = note->staff();
-      if (fretNumber == 255) {
-            fretNumber = 0;
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-            note->setGhost(true);
-            }
-      int pitch = staff->part()->instr()->stringData()->getPitch(string, fretNumber);
-      note->setFret(fretNumber);
-      note->setString(string);
-      note->setPitch(pitch);
-
-      if (tieNote) {
-            bool found = false;
-            Chord* chord     = note->chord();
-            Segment* segment = chord->segment()->prev1(Segment::Type::ChordRest);
-            int track        = note->track();
-            while (segment) {
-                  Element* e = segment->element(track);
-                  if (e) {
-                        if (e->type() == Element::Type::CHORD) {
-                              Chord* chord2 = static_cast<Chord*>(e);
-                              foreach(Note* note2, chord2->notes()) {
-                                    if (note2->string() == string) {
-                                          Tie* tie = new Tie(score);
-                                          tie->setEndNote(note);
-                                          note2->add(tie);
-                                          note->setFret(note2->fret());
-                                          note->setPitch(note2->pitch());
-                                          found = true;
-                                          break;
-                                          }
-                                    }
-                              }
-                        if (found)
-                              break;
-                        }
-                  segment = segment->prev1(Segment::Type::ChordRest);
-                  }
-            if (!found)
-                  qDebug("tied note not found, pitch %d fret %d string %d", note->pitch(), note->fret(), note->string());
-            }
-      return slur;
-      }
-
-//---------------------------------------------------------
-//   readInfo
-//---------------------------------------------------------
-
-void GuitarPro4::readInfo()
-      {
-      title        = readDelphiString();
-      subtitle     = readDelphiString();
-      artist       = readDelphiString();
-      album        = readDelphiString();
-      composer     = readDelphiString();
-      QString copyright = readDelphiString();
-      if (!copyright.isEmpty())
-            score->setMetaTag("copyright", QString("Copyright %1\nAll Rights Reserved - International Copyright Secured").arg(copyright));
-
-      transcriber  = readDelphiString();
-      instructions = readDelphiString();
-      int n = readInt();
-      for (int i = 0; i < n; ++i)
-            comments.append(readDelphiString());
-      }
-
-
-//---------------------------------------------------------
-//   read
-//---------------------------------------------------------
-
-void GuitarPro4::read(QFile* fp)
-      {
-      f      = fp;
-      curPos = 30;
-
-      readInfo();
-      readUChar();      // triplet feeling
-      readLyrics();
-
-      previousDynamic = -1;
-      previousTempo = -1;
-      int tempo  = readInt();
-      key        = readInt();
-      /*int octave =*/ readUChar();    // octave
-
-      readChannels();
-      measures = readInt();
-      staves   = readInt();
-
-      int tnumerator   = 4;
-      int tdenominator = 4;
-      for (int i = 0; i < measures; ++i) {
-            GpBar bar;
-            uchar barBits = readUChar();
-            if (barBits & 0x1)
-                  tnumerator = readUChar();
-            if (barBits & 0x2)
-                  tdenominator = readUChar();
-            if (barBits & 0x4)
-                  bar.repeatFlags = bar.repeatFlags | Repeat::START;
-            if (barBits & 0x8) {                // number of repeats
-                  bar.repeatFlags = bar.repeatFlags | Repeat::END;
-                  bar.repeats = readUChar();
-                  }
-            if (barBits & 0x10) {                      // a volta
-                  uchar voltaNumber = readUChar();
-                  while (voltaNumber > 0) {
-                        // volta information is represented as a binary number
-                        bar.volta.voltaType = GP_VOLTA_BINARY;
-                        bar.volta.voltaInfo.append(voltaNumber & 1);
-                        voltaNumber >>= 1;
-                        }
-                  }
-            if (barBits & 0x20) {
-                  bar.marker = readDelphiString();     // new section?
-                  /*int color = */ readInt();    // color?
-                  }
-            if (barBits & 0x40) {
-                  int currentKey = readUChar();
-                  /* key signatures are specified as
-                   * 1# = 1, 2# = 2, ..., 7# = 7
-                   * 1b = 255, 2b = 254, ... 7b = 249 */
-                  bar.keysig = currentKey <= 7 ? currentKey : -256+currentKey;
-                  readUChar();        // specifies major/minor mode
-                  }
-            if (barBits & 0x80)
-                  bar.barLine = BarLineType::DOUBLE;
-            bar.timesig = Fraction(tnumerator, tdenominator);
-            bars.append(bar);
-            }
-
-      //
-      // create a part for every staff
-      //
-      for (int staffIdx = 0; staffIdx < staves; ++staffIdx) {
-            Part* part = new Part(score);
-            Staff* s = new Staff(score, part, staffIdx);
-            part->insertStaff(s);
-            score->staves().push_back(s);
-            score->appendPart(part);
-            }
-
-      createMeasures();
-
-      for (int i = 0; i < staves; ++i) {
-            int tuning[GP_MAX_STRING_NUMBER];
-
-            uchar c = readUChar();   // simulations bitmask
-            if (c & 0x2) {                // 12 stringed guitar
-                  }
-            if (c & 0x4) {                // banjo track
-                  }
-            QString name = readPascalString(40);
-            int strings  = readInt();
-            if (strings <= 0 || strings > GP_MAX_STRING_NUMBER)
-                  throw GuitarProError::GP_BAD_NUMBER_OF_STRINGS ;
-            for (int j = 0; j < strings; ++j)
-                  tuning[j] = readInt();
-            for (int j = strings; j < GP_MAX_STRING_NUMBER; ++j)
-                  readInt();
-            /*int midiPort     =*/ readInt(); // - 1;
-            int midiChannel  = readInt() - 1;
-            /*int midiChannel2 =*/ readInt(); // - 1;
-            int frets        = readInt();
-
-            int capo         = readInt();
-            /*int color        =*/ readInt();
-
-            int tuning2[strings];
-            for (int k = 0; k < strings; ++k)
-                  tuning2[strings-k-1] = tuning[k];
-            StringData stringData(frets, strings, tuning2);
-            Part* part = score->staff(i)->part();
-            Instrument* instr = part->instr();
-            instr->setStringData(stringData);
-            part->setPartName(name);
-            instr->setTranspose(Interval(capo));
-            part->setLongName(name);
-
-            //
-            // determine clef
-            //
-            Staff* staff = score->staff(i);
-            int patch = channelDefaults[midiChannel].patch;
-            ClefType clefId = ClefType::G;
-            if (midiChannel == GP_DEFAULT_PERCUSSION_CHANNEL) {
-                  clefId = ClefType::PERC;
-                  instr->setUseDrumset(true);
-                        staff->setStaffType(StaffType::preset(StaffTypes::PERC_DEFAULT));
-                  }
-            else if (patch >= 24 && patch < 32)
-                  clefId = ClefType::G3;
-            else if (patch >= 32 && patch < 40)
-                  clefId = ClefType::F8;
-            Measure* measure = score->firstMeasure();
-            Clef* clef = new Clef(score);
-            clef->setClefType(clefId);
-            clef->setTrack(i * VOICES);
-            Segment* segment = measure->getSegment(Segment::Type::Clef, 0);
-            segment->add(clef);
-
-
-            Channel& ch = instr->channel(0);
-            if (midiChannel == GP_DEFAULT_PERCUSSION_CHANNEL) {
-                  ch.program = 0;
-                  ch.bank    = 128;
-                  }
-            else {
-                  ch.program = patch;
-                  ch.bank    = 0;
-                  }
-            ch.volume  = channelDefaults[midiChannel].volume;
-            ch.pan     = channelDefaults[midiChannel].pan;
-            ch.chorus  = channelDefaults[midiChannel].chorus;
-            ch.reverb  = channelDefaults[midiChannel].reverb;
-            // missing: phase, tremolo
-            ch.updateInitList();
-            }
-
-      slurs = new Slur*[staves];
-      for (int i = 0; i < staves; ++i)
-            slurs[i] = 0;
-      Measure* measure = score->firstMeasure();
-      for (int bar = 0; bar < measures; ++bar, measure = measure->nextMeasure()) {
-            const GpBar& gpbar = bars[bar];
-
-            if (!gpbar.marker.isEmpty()) {
-                  Text* s = new RehearsalMark(score);
-                  s->setText(gpbar.marker.trimmed());
-                  s->setTrack(0);
-                  Segment* segment = measure->getSegment(Segment::Type::ChordRest, measure->tick());
-                  segment->add(s);
-                  }
-
-            Tuplet* tuplets[staves];
-            for (int staffIdx = 0; staffIdx < staves; ++staffIdx)
-                  tuplets[staffIdx] = 0;
-
-            for (int staffIdx = 0; staffIdx < staves; ++staffIdx) {
-                  int track = staffIdx * VOICES;
-                  int tick  = measure->tick();
-                  int beats = readInt();
-                  for (int beat = 0; beat < beats; ++beat) {
-                        uchar beatBits = readUChar();
-                        bool dotted = beatBits & 0x1;
-                        int pause = -1;
-                        if (beatBits & 0x40)
-                              pause = readUChar();
-                        int len = readChar();
-                        int tuple = 0;
-                        if (beatBits & 0x20)
-                              tuple = readInt();
-                        Segment* segment = measure->getSegment(Segment::Type::ChordRest, tick);
-                        if (beatBits & 0x2){
-                              int numStrings = score->staff(staffIdx)->part()->instr()->stringData()->strings();
-                              int header = readUChar();
-                              QString name;
-                              if ((header & 1) == 0) {
-                                    name = readDelphiString();
-                                    readChord(segment, track, numStrings, name, false);
-                                    }
-                              else  {
-                                    skip(16);
-                                    name = readPascalString(21);
-                                    skip(4);
-                                    readChord(segment, track, numStrings, name, true);
-                                    skip(32);
-                                    }
-                              }
-                        Lyrics* lyrics = 0;
-                        if (beatBits & 0x4) {
-                              lyrics = new Lyrics(score);
-                              lyrics->setText(readDelphiString());
-                              }
-                        int beatEffects = 0;
-                        if (beatBits & 0x8)
-                              beatEffects = readBeatEffects(track, segment);
-                        if (beatBits & 0x10)
-                              readMixChange(measure);
-                        int strings = readUChar();   // used strings mask
-                        Fraction l  = len2fraction(len);
-
-                        // Some beat effects could add a Chord before this
-                        ChordRest* cr = segment->cr(track);
-
-                        if (strings == 0) {
-                              if(segment->cr(track)){
-                                    segment->remove(segment->cr(track));
-                                    delete cr;
-                                    cr = 0;
-                                    }
-                              cr = new Rest(score);
-                              }
-                        else {
-                              if(!segment->cr(track)) {
-                                    cr = new Chord(score);
-                                    Chord* chord = static_cast<Chord*>(cr);
-                                    applyBeatEffects(chord, beatEffects);
-                                    }
-                              }
-
-                        cr->setTrack(track);
-                        if (lyrics)
-                              cr->add(lyrics);
-
-                        TDuration d(l);
-                        d.setDots(dotted ? 1 : 0);
-
-                        if (dotted)
-                              l = l + (l/2);
-
-                        if (tuple) {
-                              Tuplet* tuplet = tuplets[staffIdx];
-                              if ((tuplet == 0) || (tuplet->elementsDuration() == tuplet->baseLen().fraction() * tuplet->ratio().numerator())) {
-                                    tuplet = new Tuplet(score);
-                                    tuplet->setTrack(cr->track());
-                                    tuplets[staffIdx] = tuplet;
-                                    setTuplet(tuplet, tuple);
-                                    tuplet->setParent(measure);
-                                    }
-                              tuplet->setTrack(track);
-                              tuplet->setBaseLen(l);
-                              cr->setTuplet(tuplet);
-                              tuplet->add(cr);
-                              }
-
-
-                        cr->setDuration(l);
-                        if (cr->type() == Element::Type::REST && (pause == 0 || l == measure->len()))
-                              cr->setDurationType(TDuration::DurationType::V_MEASURE);
-                        else
-                              cr->setDurationType(d);
-                        if(!segment->cr(track))
-                              segment->add(cr);
-                        Staff* staff = cr->staff();
-                        int numStrings = staff->part()->instr()->stringData()->strings();
-                        bool hasSlur = false;
-                        for (int i = 6; i >= 0; --i) {
-                              if (strings & (1 << i) && ((6-i) < numStrings)) {
-                                    Note* note = new Note(score);
-                                    static_cast<Chord*>(cr)->add(note);
-                                    hasSlur = readNote(6-i, note);
-                                    note->setTpcFromPitch();
-                                    }
-                              }
-
-                        if (hasSlur && (slurs[staffIdx] == 0)) {
-                              Slur* slur = new Slur(score);
-                              slur->setTrack(track);
-                              slur->setTrack2(track);
-                              slur->setTick(cr->tick());
-                              slur->setTick2(cr->tick());
-                              slurs[staffIdx] = slur;
-                              score->addElement(slur);
-                              }
-                        else if (slurs[staffIdx] && !hasSlur) {
-                              // TODO: check slur
-                              Slur* s = slurs[staffIdx];
-                              slurs[staffIdx] = 0;
-                              s->setTick2(cr->tick());
-                              s->setTrack2(cr->track());
-                              }
-                        else if (slurs[staffIdx] && hasSlur) {
-                              }
-                        restsForEmptyBeats(segment, measure, cr, l, track, tick);
-                        createSlur(hasSlur, staffIdx, cr);
-                        tick += cr->actualTicks();
-                        }
-                  }
-            }
-      setTempo(tempo, score->firstMeasure());
-      }
-
-//---------------------------------------------------------
-//   readInfo
-//---------------------------------------------------------
-
-void GuitarPro5::readInfo()
-      {
-      title        = readDelphiString();
-      subtitle     = readDelphiString();
-      artist       = readDelphiString();
-      album        = readDelphiString();
-      composer     = readDelphiString();
-      readDelphiString();
-      QString copyright = readDelphiString();
-      if (!copyright.isEmpty())
-            score->setMetaTag("copyright", QString("Copyright %1\nAll Rights Reserved - International Copyright Secured").arg(copyright));
-
-      transcriber  = readDelphiString();
-      instructions = readDelphiString();
-      int n = readInt();
-      for (int i = 0; i < n; ++i)
-            comments.append(readDelphiString());
-      }
-
-//---------------------------------------------------------
-//   readNoteEffects
-//---------------------------------------------------------
-
-bool GuitarPro5::readNoteEffects(Note* note)
-      {
-      uchar modMask1 = readUChar();
-      uchar modMask2 = readUChar();
-      bool slur = false;
-      if (modMask1 & 0x1)
-            readBend(note);
-      if (modMask1 & 0x2)         // hammer on / pull off
-            slur = true;
-      if (modMask1 & 0x8) {         // let ring
-            }
-      if (modMask1 & 0x10) {
-            int fret       = readUChar();       // grace fret
-            int dynamic    = readUChar();       // grace dynamic
-            int transition = readUChar();       // grace transition
-            int duration   = readUChar();       // grace duration
-            int gflags     = readUChar();
-
-            int grace_len = MScore::division/8;
-            NoteType note_type =  NoteType::ACCIACCATURA;
-
-            if(gflags & 0x02) //on beat
-                  note_type = NoteType::APPOGGIATURA;
-
-            if (duration == 1)
-                  grace_len = MScore::division/8; //32th
-            else if (duration == 2)
-                  grace_len = MScore::division/6; //24th
-            else if (duration == 3)
-                  grace_len = MScore::division/4; //16th
-
-            Note* gn = new Note(score);
-
-            if (gflags & 0x01) {
-                  gn->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-                  gn->setGhost(true);
-                  }
-            gn->setFret(fret);
-            gn->setString(note->string());
-            int grace_pitch = note->staff()->part()->instr()->stringData()->getPitch(note->string(), fret);
-            gn->setPitch(grace_pitch);
-            gn->setTpcFromPitch();
-
-            Chord* gc = new Chord(score);
-            gc->add(gn);
-            note->chord()->add(gc);
-
-            // TODO: Add dynamic. Dynamic now can be added only to a segment, not directly to a grace note
-            addDynamic(gn, dynamic);
-
-            TDuration d;
-            d.setVal(grace_len);
-            if(grace_len == MScore::division/6)
-                  d.setDots(1);
-            gc->setDurationType(d);
-            gc->setDuration(d.fraction());
-            gc->setNoteType(note_type);
-            gc->setMag(note->chord()->staff()->mag() * score->styleD(StyleIdx::graceNoteMag));
-            if (transition == 0) {
-                  // no transition
-                  }
-            else if(transition == 1){
-                  //TODO: Add a 'slide' guitar effect when implemented
-                  }
-            else if (transition == 2 && note->fret()>=0 && note->fret()<=255 && note->fret()!=gn->fret()) {
-                  QList<PitchValue> points;
-                  points.append(PitchValue(0,0, false));
-                  points.append(PitchValue(60,(note->fret()-gn->fret())*100, false));
-
-                  Bend* b = new Bend(note->score());
-                  b->setPoints(points);
-                  b->setTrack(gn->track());
-                  gn->add(b);
-                  }
-             else if (transition == 3) {
-                   // TODO:
-                   //     major: replace with a 'hammer-on' guitar effect when implemented
-                   //     minor: make slurs for parts
-
-                   ChordRest* cr1 = static_cast<Chord*>(gc);
-                   ChordRest* cr2 = static_cast<Chord*>(note->chord());
-
-                   Slur* slur = new Slur(score);
-                   slur->setStartChord(static_cast<Chord*>(cr1));
-                   slur->setEndChord(static_cast<Chord*>(cr2));
-                   slur->setTick(cr1->tick());
-                   slur->setTick2(cr2->tick());
-                   slur->setTrack(cr1->track());
-                   slur->setTrack2(cr2->track());
-                   score->addElement(slur);
-                   }
-            }
-      if (modMask2 & 0x1) {   // staccato - palm mute
-            Chord* chord = note->chord();
-            Articulation* a = new Articulation(chord->score());
-            a->setArticulationType(ArticulationType::Staccato);
-            chord->add(a);
-            }
-      if (modMask2 & 0x2) {   // palm mute - mute the whole column
-            }
-      if (modMask2 & 0x4) {    // tremolo picking length
-            int tremoloDivision = readUChar();
-            Chord* chord = note->chord();
-            Tremolo* t = new Tremolo(chord->score());
-            if (tremoloDivision == 1) {
-                  t->setTremoloType(TremoloType::R8);
-                  chord->add(t);
-                  }
-            else if (tremoloDivision == 2) {
-                  t->setTremoloType(TremoloType::R16);
-                  chord->add(t);
-                  }
-            else if (tremoloDivision == 3) {
-                  t->setTremoloType(TremoloType::R32);
-                  chord->add(t);
-                  }
-            else
-                  qDebug("Unknown tremolo value");
-      }
-      if (modMask2 & 0x8)
-            readUChar();      // slide kind
-      if (modMask2 & 0x10)
-            readArtificialHarmonic();
-      if (modMask2 & 0x20) {
-            readUChar();      // trill fret
-            int period = readUChar();      // trill length
-
-            // add the trill articulation to the note
-            Articulation* art = new Articulation(note->score());
-            art->setArticulationType(ArticulationType::Trill);
-            if (!note->score()->addArticulation(note, art))
-                  delete art;
-
-            switch(period) {
-                  case 1:           // 16
-                        break;
-                  case 2:           // 32
-                        break;
-                  case 3:           // 64
-                        break;
-                  default:
-                        qDebug("unknown trill period %d", period);
-                        break;
-                  }
-            }
-      return slur;
-      }
-
-//---------------------------------------------------------
-//   readNote
-//---------------------------------------------------------
-
-bool GuitarPro5::readNote(int string, Note* note)
-      {
-      uchar noteBits = readUChar();
-      //
-      // noteBits:
-      //    7 - Right hand or left hand fingering;
-      //    6 - Accentuated note
-      //    5 - Note type (rest, empty note, normal note);
-      //    4 - note dynamic;
-      //    3 - Presence of effects linked to the note;
-      //    2 - Ghost note;
-      //    1 - Dotted note;  ?
-      //    0 - Time-independent duration
-
-      if (noteBits & 0x04) {
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-            note->setGhost(true);
-            }
-
-      bool tieNote = false;
-      if (noteBits & 0x20) {
-            uchar noteType = readUChar();
-            if (noteType == 1) {
-                  }
-            else if (noteType == 2) {
-                  tieNote = true;
-                  }
-            else if (noteType == 3) {                 // dead notes
-                  note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-                  note->setGhost(true);
-                  }
-            else
-                  qDebug("unknown note type: %d", noteType);
-            }
-
-      if (noteBits & 0x10) {          // velocity
-            int d = readChar();
-            if (previousDynamic != d) {
-                  previousDynamic = d;
-                  addDynamic(note, d);
-                  }
-            }
-
-      int fretNumber = 0;
-      if (noteBits & 0x20)
-            fretNumber = readChar();
-
-      if (noteBits & 0x80) {              // fingering
-            int a = readUChar();
-            int b = readUChar();
-            qDebug("   Fingering=========%d %d", a, b);
-            }
-      if (noteBits & 0x1)
-            skip(8);
-
-      // check if a note is supposed to be accented, and give it the marcato type
-      if (noteBits & 0x02) {
-            Articulation* art = new Articulation(note->score());
-            art->setArticulationType(ArticulationType::Marcato);
-            if (!note->score()->addArticulation(note, art))
-                  delete art;
-      }
-
-      // check if a note is supposed to be accented, and give it the sforzato type
-      if (noteBits & 0x40) {
-            Articulation* art = new Articulation(note->score());
-            art->setArticulationType(ArticulationType::Sforzatoaccent);
-            if (!note->score()->addArticulation(note, art))
-                  delete art;
-            }
-
-      readUChar(); //skip
-
-      Staff* staff = note->staff();
-      if (fretNumber == 255) {
-            fretNumber = 0;
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-            note->setGhost(true);
-            }
-      int pitch = staff->part()->instr()->stringData()->getPitch(string, fretNumber);
-      note->setFret(fretNumber);
-      note->setString(string);
-      note->setPitch(pitch);
-
-      // This function uses string and fret number, so it should be set before this
-      bool slur = false;
-      if (noteBits & 0x8)
-            slur = readNoteEffects(note);
-
-      if (tieNote) {
-            bool found = false;
-            Chord* chord     = note->chord();
-            Segment* segment = chord->segment()->prev1(Segment::Type::ChordRest);
-            int track        = note->track();
-            while (segment) {
-                  Element* e = segment->element(track);
-                  if (e) {
-                        if (e->type() == Element::Type::CHORD) {
-                              Chord* chord2 = static_cast<Chord*>(e);
-                              foreach(Note* note2, chord2->notes()) {
-                                    if (note2->string() == string) {
-                                          Tie* tie = new Tie(score);
-                                          tie->setEndNote(note);
-                                          note2->add(tie);
-                                          note->setFret(note2->fret());
-                                          note->setPitch(note2->pitch());
-                                          found = true;
-                                          break;
-                                          }
-                                    }
-                              }
-                        if (found)
-                              break;
-                        }
-                  segment = segment->prev1(Segment::Type::ChordRest);
-                  }
-            if (!found)
-                  qDebug("tied note not found, pitch %d fret %d string %d", note->pitch(), note->fret(), note->string());
-            }
-      return slur;
-      }
-
-//---------------------------------------------------------
-//   readArtificialHarmonic
-//---------------------------------------------------------
-
-void GuitarPro5::readArtificialHarmonic()
-      {
-      int type = readChar();
-      switch(type) {
-            case 1:           // natural
-                  break;
-            case 2:           // artificial
-                  skip(3);
-                  break;
-            case 3:           // tapped
-                  skip(1);
-                  break;
-            case 4:           // pinch
-                  break;
-            case 5:           // semi
-                  break;
-            }
-      }
-
-//---------------------------------------------------------
 //   readTremoloBar
 //---------------------------------------------------------
 
@@ -2780,8 +1851,6 @@ void GuitarPro::readTremoloBar(int /*track*/, Segment* /*segment*/)
       /*int a4 =*/ readChar();
       /*int a5 =*/ readChar();
       int n  =  readInt();
-
-
       QList<PitchValue> points;
       for (int i = 0; i < n; ++i) {
             int time    = readInt();
@@ -2795,1488 +1864,6 @@ void GuitarPro::readTremoloBar(int /*track*/, Segment* /*segment*/)
       b->setTrack(track);
       segment->add(b);*/
       }
-
-//---------------------------------------------------------
-//   readBeatEffects
-//---------------------------------------------------------
-
-int GuitarPro5::readBeatEffects(int track, Segment* segment)
-      {
-      int effects = 0;
-
-      uchar fxBits1 = readUChar();
-      uchar fxBits2 = readUChar();
-      if (fxBits1 & 0x10) {
-            effects = 4; // fade in
-            }
-      if (fxBits1 & 0x20) {
-            effects = readUChar();
-            // 1 - tapping
-            // 2 - slapping
-            // 3 - popping
-            }
-      if (fxBits2 & 0x04)
-            readTremoloBar(track, segment);       // readBend();
-      if (fxBits1 & 0x40) {
-                  int strokeup = readUChar();            // up stroke length
-                  int strokedown = readUChar();            // down stroke length
-
-                  Arpeggio* a = new Arpeggio(score);
-                  if( strokeup > 0 ) {
-                        a->setArpeggioType(ArpeggioType::UP);
-                        }
-                  else if( strokedown > 0 ) {
-                        a->setArpeggioType(ArpeggioType::DOWN);
-                        }
-                  else {
-                        delete a;
-                        a = 0;
-                        }
-
-                  if(a) {
-                        ChordRest* cr = new Chord(score);
-                        cr->setTrack(track);
-                        cr->add(a);
-                        segment->add(cr);
-                        }
-                  }
-      if (fxBits2 & 0x02) {
-            int a = readChar();            // stroke pick direction
-            qDebug("  0x02: 0x%02x", a);
-            }
-      return effects;
-      }
-
-//---------------------------------------------------------
-//   readPageSetup
-//---------------------------------------------------------
-
-void GuitarPro5::readPageSetup()
-      {
-      skip(version > 500 ? 49 : 30);
-      for (int i = 0; i < 11; ++i) {
-            skip(4);
-            readBytePascalString();
-            }
-      }
-
-//---------------------------------------------------------
-//   readBeat
-//---------------------------------------------------------
-
-int GuitarPro5::readBeat(int tick, int voice, Measure* measure, int staffIdx, Tuplet** tuplets)
-      {
-// qDebug("readBeat tick %d voice %d staffIdx %d", tick, voice, staffIdx);
-      uchar beatBits = readUChar();
-      bool dotted    = beatBits & 0x1;
-
-      int pause = -1;
-      if (beatBits & 0x40)
-            pause = readUChar();
-
-      int track = staffIdx * VOICES + voice;
-
-      // readDuration
-      int len   = readChar();
-      int tuple = 0;
-      if (beatBits & 0x20)
-            tuple = readInt();
-
-      Segment* segment = measure->getSegment(Segment::Type::ChordRest, tick);
-      if (beatBits & 0x2) {
-            int numStrings = score->staff(staffIdx)->part()->instr()->stringData()->strings();
-            skip(17);
-            QString name = readPascalString(21);
-            skip(4);
-            // no header to be read in the GP5 format - default to true.
-            readChord(segment, staffIdx * VOICES, numStrings, name, true);
-            skip(32);
-            }
-      Lyrics* lyrics = 0;
-      if (beatBits & 0x4) {
-            QString txt = readDelphiString();
-            lyrics = new Lyrics(score);
-            lyrics->setText(txt);
-            }
-
-      int beatEffects = 0;
-      if (beatBits & 0x8)
-            beatEffects = readBeatEffects(track, segment);
-      if (beatBits & 0x10)
-            readMixChange(measure);
-
-      int strings = readUChar();   // used strings mask
-
-      Fraction l    = len2fraction(len);
-
-      // Some beat effects could add a Chord before this
-      ChordRest* cr = segment->cr(track);
-      if (voice != 0 && pause == 0 && strings == 0)
-            cr = 0;
-      else {
-            if (strings == 0) {
-                  if (cr) {
-                        segment->remove(cr);
-                        delete cr;
-                        cr = 0;
-                        }
-                  cr = new Rest(score);
-                  }
-            else  {
-                  if (!cr)
-                        cr = new Chord(score);
-                  }
-            cr->setTrack(track);
-
-            TDuration d(l);
-            d.setDots(dotted ? 1 : 0);
-
-            if (dotted)
-                  l = l + (l/2);
-
-            if (tuple) {
-                  Tuplet* tuplet = tuplets[staffIdx * 2 + voice];
-                  if ((tuplet == 0) || (tuplet->elementsDuration() == tuplet->baseLen().fraction() * tuplet->ratio().numerator())) {
-                        tuplet = new Tuplet(score);
-                        // int track = staffIdx * 2 + voice;
-                        tuplets[staffIdx * 2 + voice] = tuplet;
-                        tuplet->setTrack(cr->track());
-                        setTuplet(tuplet, tuple);
-                        tuplet->setParent(measure);
-                        }
-                  tuplet->setTrack(cr->track());
-                  tuplet->setBaseLen(l);
-                  cr->setTuplet(tuplet);
-                  tuplet->add(cr);
-                  }
-
-            cr->setDuration(l);
-            if (cr->type() == Element::Type::REST && pause == 0)
-                  cr->setDurationType(TDuration::DurationType::V_MEASURE);
-            else
-                  cr->setDurationType(d);
-
-            if(!segment->cr(track))
-                  segment->add(cr);
-
-            Staff* staff = cr->staff();
-            int numStrings = staff->part()->instr()->stringData()->strings();
-            bool hasSlur = false;
-            for (int i = 6; i >= 0; --i) {
-                  if (strings & (1 << i) && ((6-i) < numStrings)) {
-                        Note* note = new Note(score);
-                        static_cast<Chord*>(cr)->add(note);
-                        hasSlur = readNote(6-i, note);
-                        note->setTpcFromPitch();
-                        }
-                  }
-            createSlur(hasSlur, staffIdx, cr);
-
-            if (lyrics)
-                  cr->add(lyrics);
-            }
-      int rr = readChar();
-      if (cr && (cr->type() == Element::Type::CHORD)) {
-            Chord* chord = static_cast<Chord*>(cr);
-            applyBeatEffects(chord, beatEffects);
-            if (rr == 0x2)
-                  chord->setStemDirection(MScore::Direction::DOWN);
-            else if (rr == 0xa)
-                  chord->setStemDirection(MScore::Direction::UP);
-            else {
-                  ; // qDebug("  1beat read 0x%02x", rr);
-                  }
-            }
-      int r = readChar();
-      if (r & 0x8) {
-            int rrr = readChar();
-qDebug("  3beat read 0x%02x", rrr);
-            }
-      restsForEmptyBeats(segment, measure, cr, l, track, tick);
-      return cr ? cr->actualTicks() : measure->ticks();
-      }
-
-//---------------------------------------------------------
-//   readMeasure
-//---------------------------------------------------------
-
-void GuitarPro5::readMeasure(Measure* measure, int staffIdx, Tuplet** tuplets)
-      {
-      for (int voice = 0; voice < 2; ++voice) {
-            int tick = measure->tick();
-            int beats = readInt();
-            for (int beat = 0; beat < beats; ++beat) {
-                  tick += readBeat(tick, voice, measure, staffIdx, tuplets);
-                  }
-            }
-      }
-
-//---------------------------------------------------------
-//   readMixChange
-//---------------------------------------------------------
-
-void GuitarPro5::readMixChange(Measure* measure)
-      {
-      /*char patch   =*/ readChar();
-      skip(16);
-      char volume  = readChar();
-      char pan     = readChar();
-      char chorus  = readChar();
-      char reverb  = readChar();
-      char phase   = readChar();
-      char tremolo = readChar();
-      readDelphiString();                 // tempo name
-
-      int tempo = readInt();
-
-      if (volume >= 0)
-            readChar();
-      if (pan >= 0)
-            readChar();
-      if (chorus >= 0)
-            readChar();
-      if (reverb >= 0)
-            readChar();
-      if (phase >= 0)
-            readChar();
-      if (tremolo >= 0)
-            readChar();
-      if (tempo >= 0) {
-            if (tempo != previousTempo) {
-                  previousTempo = tempo;
-                  setTempo(tempo, measure);
-                  }
-            readChar();
-            if (version > 500)
-                  readChar();
-            }
-      readChar();
-      skip(1);
-      if (version > 500) {
-            readDelphiString();
-            readDelphiString();
-            }
-      }
-
-
-//---------------------------------------------------------
-//   readTracks
-//---------------------------------------------------------
-
-void GuitarPro5::readTracks()
-      {
-      for (int i = 0; i < staves; ++i) {
-            int tuning[GP_MAX_STRING_NUMBER];
-            Staff* staff = score->staff(i);
-            Part* part = staff->part();
-
-            uchar c = readUChar();   // simulations bitmask
-            if (c & 0x2) {                // 12 stringed guitar
-                  }
-            if (c & 0x4) {                // banjo track
-                  }
-            if (i == 0 || version == 500)
-                  skip(1);
-            QString name = readPascalString(40);
-
-            int strings  = readInt();
-            if (strings <= 0 || strings > GP_MAX_STRING_NUMBER)
-                  throw GuitarProError::GP_BAD_NUMBER_OF_STRINGS ;
-            for (int j = 0; j < strings; ++j) {
-                  tuning[j] = readInt();
-                  }
-            for (int j = strings; j < GP_MAX_STRING_NUMBER; ++j)
-                  readInt();
-            /*int midiPort     =*/ readInt();   // -1
-            int midiChannel  = readInt() - 1;
-            /*int midiChannel2 =*/ readInt();   // -1
-
-            int frets        = readInt();
-            int capo         = readInt();
-            /*int color        =*/ readInt();
-
-            skip(version > 500 ? 49 : 44);
-            if (version > 500) {
-                  readDelphiString();
-                  readDelphiString();
-                  }
-
-            int tuning2[strings];
-            for (int k = 0; k < strings; ++k)
-                  tuning2[strings-k-1] = tuning[k];
-            StringData stringData(frets, strings, tuning2);
-            Instrument* instr = part->instr();
-            instr->setStringData(stringData);
-            part->setPartName(name);
-            part->setLongName(name);
-            instr->setTranspose(Interval(capo));
-
-            //
-            // determine clef
-            //
-            int patch = channelDefaults[midiChannel].patch;
-            ClefType clefId = ClefType::G;
-            if (midiChannel == GP_DEFAULT_PERCUSSION_CHANNEL) {
-                  clefId = ClefType::PERC;
-                  instr->setUseDrumset(true);
-                        staff->setStaffType(StaffType::preset(StaffTypes::PERC_DEFAULT));
-                  }
-            else if (patch >= 24 && patch < 32)
-                  clefId = ClefType::G3;
-            else if (patch >= 32 && patch < 40)
-                  clefId = ClefType::F8;
-            Measure* measure = score->firstMeasure();
-            Clef* clef = new Clef(score);
-            clef->setClefType(clefId);
-            clef->setTrack(i * VOICES);
-            Segment* segment = measure->getSegment(Segment::Type::Clef, 0);
-            segment->add(clef);
-
-            Channel& ch = instr->channel(0);
-            if (midiChannel == GP_DEFAULT_PERCUSSION_CHANNEL) {
-                  ch.program = 0;
-                  ch.bank    = 128;
-                  }
-            else {
-                  ch.program = patch;
-                  ch.bank    = 0;
-                  }
-            ch.volume  = channelDefaults[midiChannel].volume;
-            ch.pan     = channelDefaults[midiChannel].pan;
-            ch.chorus  = channelDefaults[midiChannel].chorus;
-            ch.reverb  = channelDefaults[midiChannel].reverb;
-            // missing: phase, tremolo
-            ch.updateInitList();
-            }
-      skip(version == 500 ? 2 : 1);
-      }
-
-//---------------------------------------------------------
-//   readMeasures
-//---------------------------------------------------------
-
-void GuitarPro5::readMeasures()
-      {
-      Measure* measure = score->firstMeasure();
-      for (int bar = 0; bar < measures; ++bar, measure = measure->nextMeasure()) {
-            const GpBar& gpbar = bars[bar];
-
-            if (!gpbar.marker.isEmpty()) {
-                  Text* s = new RehearsalMark(score);
-                  s->setText(gpbar.marker.trimmed());
-                  s->setTrack(0);
-                  Segment* segment = measure->getSegment(Segment::Type::ChordRest, measure->tick());
-                  segment->add(s);
-                  }
-
-            Tuplet* tuplets[staves * 2];     // two voices
-            for (int track = 0; track < staves*2; ++track)
-                  tuplets[track] = 0;
-
-            for (int staffIdx = 0; staffIdx < staves; ++staffIdx) {
-                  readMeasure(measure, staffIdx, tuplets);
-                  if (!(((bar == (measures-1)) && (staffIdx == (staves-1))))) {
-                        /*int a = */  readChar();
-                        // qDebug("    ======skip %02x", a);
-                        }
-                  }
-            }
-      }
-
-//---------------------------------------------------------
-//   read
-//---------------------------------------------------------
-
-void GuitarPro5::read(QFile* fp)
-      {
-      f = fp;
-      readInfo();
-      readLyrics();
-      readPageSetup();
-
-      previousDynamic = -1;
-      previousTempo = -1;
-      int tempo = readInt();
-      if (version > 500)
-            skip(1);
-
-      key    = readInt();
-      /* int octave =*/ readChar();    // octave
-
-      readChannels();
-      skip(42);
-
-      measures = readInt();
-      staves  = readInt();
-
-      slurs = new Slur*[staves];
-      for (int i = 0; i < staves; ++i)
-            slurs[i] = 0;
-
-      int tnumerator   = 4;
-      int tdenominator = 4;
-      for (int i = 0; i < measures; ++i) {
-            if (i > 0)
-                  skip(1);
-            GpBar bar;
-            uchar barBits = readUChar();
-            if (barBits & 0x1)
-                  tnumerator = readUChar();
-            if (barBits & 0x2)
-                  tdenominator = readUChar();
-            if (barBits & 0x4)
-                  bar.repeatFlags = bar.repeatFlags | Repeat::START;
-            if (barBits & 0x8) {                // number of repeats
-                  bar.repeatFlags = bar.repeatFlags | Repeat::END;
-                  bar.repeats = readUChar();
-                  }
-            if (barBits & 0x20) {
-                  bar.marker = readDelphiString();     // new section?
-                  /*int color =*/ readInt();    // color?
-                  }
-            if (barBits & 0x10) {                      // a volta
-                  uchar voltaNumber = readUChar();
-                  while (voltaNumber > 0) {
-                        // voltas are represented as flags
-                        bar.volta.voltaType = GP_VOLTA_FLAGS;
-                        bar.volta.voltaInfo.append(voltaNumber & 1);
-                        voltaNumber >>= 1;
-                        }
-                  }
-            if (barBits & 0x40) {
-                  int currentKey = readUChar();
-                  /* key signatures are specified as
-                   * 1# = 1, 2# = 2, ..., 7# = 7
-                   * 1b = 255, 2b = 254, ... 7b = 249 */
-                  bar.keysig = currentKey <= 7 ? currentKey : -256+currentKey;
-                  readUChar();        // specified major/minor mode
-                  }
-            if (barBits & 0x80)
-                  bar.barLine = BarLineType::DOUBLE;
-            if (barBits & 0x3)
-                  skip(4);
-            if ((barBits & 0x10) == 0)
-                  skip(1);
-            readChar();             // triple feel  (none, 8, 16)
-            bar.timesig = Fraction(tnumerator, tdenominator);
-            bars.append(bar);
-            }
-
-      //
-      // create a part for every staff
-      //
-      for (int staffIdx = 0; staffIdx < staves; ++staffIdx) {
-            Part* part = new Part(score);
-            Staff* s = new Staff(score, part, staffIdx);
-            part->insertStaff(s);
-            score->staves().push_back(s);
-            score->appendPart(part);
-            }
-
-      createMeasures();
-      readTracks();
-      readMeasures();
-      setTempo(tempo, score->firstMeasure());
-      }
-
-
-//---------------------------------------------------------
-//   readBit
-//---------------------------------------------------------
-
-int GuitarPro6::readBit() {
-      // calculate the byte index by dividing the position in bits by the bits per byte
-      int byteIndex = this->position / BITS_IN_BYTE;
-      // calculate our offset so we know how much to bit shift
-      int byteOffset = ((BITS_IN_BYTE-1) - (this->position % BITS_IN_BYTE));
-      // calculate the bit which we want to read
-      int bit = ((((*buffer)[byteIndex] & 0xff) >> byteOffset) & 0x01);
-      // increment our curent position so we know this bit has been read
-      this->position++;
-      // return the bit we calculated
-      return bit;
-      }
-
-
-//---------------------------------------------------------
-//   readBits
-//---------------------------------------------------------
-
-int GuitarPro6::readBits(int bitsToRead) {
-      int bits = 0;
-      for (int i = (bitsToRead-1); i>=0; i--) {
-            bits |= (readBit() << i);
-            }
-      return bits;
-      }
-
-//---------------------------------------------------------
-//   readBitsReversed
-//---------------------------------------------------------
-
-int GuitarPro6::readBitsReversed(int bitsToRead) {
-      int bits = 0;
-      for( int i = 0 ; i < bitsToRead ; i ++ ) {
-            bits |= (readBit() << i );
-            }
-      return bits;
-      }
-
-//---------------------------------------------------------
-//   getBytes
-//---------------------------------------------------------
-
-QByteArray GuitarPro6::getBytes(QByteArray* buffer, int offset, int length) {
-      QByteArray newBytes;
-      // compute new bytes from our buffer and return byte array
-      for (int i = 0; i < length; i++) {
-            if (buffer->length() > offset + i) {
-                  newBytes.insert(i, ((*buffer)[offset+i]));
-                  }
-            }
-      return newBytes;
-      }
-
-//---------------------------------------------------------
-//   readInteger
-//---------------------------------------------------------
-
-int GuitarPro6::readInteger(QByteArray* buffer, int offset) {
-      // assign four bytes and take them from the buffer
-      char bytes[4];
-      bytes[0] = (*buffer)[offset+0];
-      bytes[1] = (*buffer)[offset+1];
-      bytes[2] = (*buffer)[offset+2];
-      bytes[3] = (*buffer)[offset+3];
-      // increment positioning so we keep track of where we are
-      this->position+=sizeof(int)*BITS_IN_BYTE;
-      // bit shift in order to compute our integer value and return
-      return ((bytes[3] & 0xff) << 24) | ((bytes[2] & 0xff) << 16) | ((bytes[1] & 0xff) << 8) | (bytes[0] & 0xff);
-      }
-
-//---------------------------------------------------------
-//   readString
-//---------------------------------------------------------
-
-QByteArray GuitarPro6::readString(QByteArray* buffer, int offset, int length) {
-      QByteArray filename;
-      // compute the string by iterating through the buffer
-      for (int i = 0; i < length; i++) {
-            int charValue = (((*buffer)[offset + i]) & 0xff);
-            if (charValue == 0)
-                  break;
-            filename.push_back((char)charValue);
-      }
-      return filename;
-}
-
-//---------------------------------------------------------
-//   readGPX
-//---------------------------------------------------------
-
-void GuitarPro6::readGPX(QByteArray* buffer) {
-      // start by reading the file header. It will tell us if the byte array is compressed.
-      int fileHeader = readInteger(buffer, 0);
-
-      if (fileHeader == GPX_HEADER_COMPRESSED) {
-            // this is  a compressed file.
-            int length = readInteger(buffer, this->position/BITS_IN_BYTE);
-            QByteArray* bcfsBuffer = new QByteArray();
-            int positionCounter = 0;
-            while(!f->error() && (this->position/this->BITS_IN_BYTE) < length) {
-                  // read the bit indicating compression information
-                  int flag = this->readBits(1);
-
-                  if (flag) {
-                        int bits = this->readBits(4);
-                        int offs = this->readBitsReversed(bits);
-                        int size = this->readBitsReversed(bits);
-
-                        QByteArray bcfsBufferCopy = *bcfsBuffer;
-                        int pos = (bcfsBufferCopy.length() - offs );
-                        for( int i = 0; i < (size > offs ? offs : size) ; i ++ ) {
-                              bcfsBuffer->insert(positionCounter, bcfsBufferCopy[pos + i] ) ;
-                              positionCounter++;
-                              }
-                        }
-                  else  {
-                        int size = this->readBitsReversed(2);
-                        for(int i = 0; i < size; i++) {
-                              bcfsBuffer->insert(positionCounter, this->readBits(8));
-                              positionCounter++;
-                              }
-                        }
-                  }
-             // recurse on the decompressed file stored as a byte array
-             readGPX(bcfsBuffer);
-             delete bcfsBuffer;
-            }
-      else if (fileHeader == GPX_HEADER_UNCOMPRESSED) {
-            // this is an uncompressed file - strip the header off
-            *buffer = buffer->right(buffer->length()-sizeof(int));
-            int sectorSize = 0x1000;
-            int offset = 0;
-            while ((offset = (offset + sectorSize)) + 3 < buffer->length()) {
-                  int newInt = readInteger(buffer,offset);
-                  if (newInt == 2) {
-                        int indexFileName = (offset + 4);
-                        int indexFileSize = (offset + 0x8C);
-                        int indexOfBlock = (offset + 0x94);
-
-                        // create a byte array and put information about files found in it
-                        int block = 0;
-                        int blockCount = 0;
-                        QByteArray* fileBytes = new QByteArray();
-                        while((block = (readInteger(buffer, (indexOfBlock + (4 * (blockCount ++)))))) != 0 ) {
-                              fileBytes->push_back(getBytes(buffer, (offset = (block*sectorSize)), sectorSize));
-                              }
-                        // get file information and read the file
-                        int fileSize = readInteger(buffer, indexFileSize);
-                        if (fileBytes->length() >= fileSize) {
-                              QByteArray filenameBytes = readString(buffer, indexFileName, 127);
-                              char* filename = filenameBytes.data();
-                              QByteArray data = getBytes(fileBytes, 0, fileSize);
-                              parseFile(filename, &data);
-                              }
-                        delete fileBytes;
-                        }
-                  }
-            }
-}
-
-//---------------------------------------------------------
-//   parseFile
-//---------------------------------------------------------
-
-void GuitarPro6::parseFile(char* filename, QByteArray* data)
-      {
-      // test to check if we are dealing with the score
-      if (!strcmp(filename, "score.gpif")) {
-            QFile file("/tmp/score.gpif");
-            file.open(QIODevice::WriteOnly);
-            file.write(*data);
-            file.close();
-            readGpif(data);
-            }
-      }
-
-
-//---------------------------------------------------------
-//   unhandledNode
-//---------------------------------------------------------
-
-void GuitarPro6::unhandledNode(QString nodeName)
-      {
-      qDebug() << "WARNING: Discovered unhandled node name" << nodeName;
-      }
-
-//---------------------------------------------------------
-//   readScore
-//---------------------------------------------------------
-
-void GuitarPro6::readScore(QDomNode* scoreNode)
-      {
-      // loop through the score meta-info, grabbing title, artist, etc
-      QDomNode currentNode = scoreNode->firstChild();
-      while (!currentNode.isNull()) {
-            QString nodeName = currentNode.nodeName();
-            if (!nodeName.compare("Title"))
-                  title = currentNode.toElement().text();
-            else if (!nodeName.compare("Subtitle"))
-                  subtitle = currentNode.toElement().text();
-            else if (!nodeName.compare("Artist"))
-                  artist = currentNode.toElement().text();
-            else if (!nodeName.compare("Album"))
-                  album = currentNode.toElement().text();
-            else
-                  unhandledNode(nodeName);
-            currentNode = currentNode.nextSibling();
-            }
-      }
-
-//---------------------------------------------------------
-//   readMasterTracks
-//---------------------------------------------------------
-
-void GuitarPro6::readMasterTracks(QDomNode* masterTrack)
-      {
-      // inspects MasterTrack, gives information applying to start of score such as tempo
-      QDomNode currentNode = masterTrack->firstChild();
-      while (!currentNode.isNull()) {
-            QString nodeName = currentNode.nodeName();
-            if (!nodeName.compare("Automations")) {
-                  QDomNode currentAutomation = currentNode.firstChild();
-                  while (!currentAutomation.isNull()) {
-                        if (!currentAutomation.nodeName().compare("Automation")) {
-                              if (!currentAutomation.firstChild().nodeName().compare("Tempo"))
-                                    tempo = currentAutomation.lastChild().toElement().text().toInt();
-                              else
-                                    unhandledNode(currentAutomation.nodeName());
-                              }
-                        else
-                              unhandledNode(currentAutomation.nodeName());
-                        currentAutomation = currentAutomation.nextSibling();
-                        }
-                  }
-            currentNode = currentNode.nextSibling();
-            }
-      }
-
-//---------------------------------------------------------
-//   readChord
-//---------------------------------------------------------
-
-void GuitarPro6::readChord(QDomNode* diagram, int track)
-      {
-      // initialise a new fret diagram for our current track
-      FretDiagram* fretDiagram = new FretDiagram(score);
-      fretDiagram->setTrack(track);
-
-      // get the identifier to set as the domain in the map
-      int id = diagram->attributes().namedItem("id").toAttr().value().toInt();
-      QDomNode diagramNode = diagram->firstChild();
-
-      // set the number of strings on this part
-      int stringCount = diagramNode.attributes().namedItem("stringCount").toAttr().value().toInt();
-      fretDiagram->setStrings(stringCount);
-
-      // set the fret offset
-      int baseFret = diagramNode.attributes().namedItem("baseFret").toAttr().value().toInt();
-      fretDiagram->setOffset(baseFret);
-
-      QDomNode diagramEntity = diagramNode.firstChild();
-      int counter = 0;
-      while (!diagramEntity.isNull()) {
-            QString nodeName = diagramEntity.nodeName();
-            // new fret
-            if (!nodeName.compare("Fret")) {
-                  // get the string and fret numbers from the arguments to the node as integers
-                  int string = diagramEntity.attributes().namedItem("string").toAttr().value().toInt();
-                  int fret = diagramEntity.attributes().namedItem("fret").toAttr().value().toInt();
-
-                  // if there are unspecified string values, add the X marker to that string
-                  while (counter < string) {
-                        fretDiagram->setMarker(counter, 'X');
-                        counter++;
-                        }
-
-                  // look at the specified string/fret and add to diagram
-                  if (fret == 0) {
-                        fretDiagram->setMarker(string, '0');
-                        counter++;
-                        }
-                  else  {
-                        qDebug("Setting fret (string=%d,fret=%d)", string, fret-baseFret);
-                        fretDiagram->setDot(string, fret);
-                        counter++;
-                        }
-                  }
-            // move to the next string/fret specification
-            diagramEntity = diagramEntity.nextSibling();
-            }
-
-      // mark any missing strings as 'X'
-      while (counter < stringCount) {
-            fretDiagram->setMarker(counter, 'X');
-            counter++;
-            }
-
-      // insert the fret diagram into the map of diagrams
-      fretDiagrams.insert(id, fretDiagram);
-      }
-
-//---------------------------------------------------------
-//   readTracks
-//---------------------------------------------------------
-
-void GuitarPro6::readTracks(QDomNode* track)
-      {
-      QDomNode nextTrack = track->firstChild();
-      int trackCounter = 0;
-      while (!nextTrack.isNull()) {
-            QDomNode currentNode = nextTrack.firstChild();
-            Part* part = new Part(score);
-            Staff* s = new Staff(score, part, trackCounter);
-            while (!currentNode.isNull()) {
-                  QString nodeName = currentNode.nodeName();
-                  if (nodeName == "Name")
-                        part->setLongName(currentNode.toElement().text());
-                  else if (nodeName == "ShortName")
-                        part->setPartName(currentNode.toElement().text());
-                  else if (nodeName == "Instrument") {
-                        QString ref = currentNode.attributes().namedItem("ref").toAttr().value();
-                        Instrument* instr = new Instrument;
-                        // use an array as a map instead?
-                        if (!ref.compare("e-gtr6")) {
-                              *instr = instr->fromTemplate(Ms::searchTemplate("electric-guitar"));
-                              part->setInstrument(*instr, 0);
-                              }
-                        else if (!ref.compare("tnr-s")) {
-                              *instr = instr->fromTemplate(Ms::searchTemplate("voice"));
-                              part->setInstrument(*instr, 0);
-                              }
-                        else if (!ref.compare("s-gtr6")) {
-                              *instr = instr->fromTemplate(Ms::searchTemplate("guitar-steel"));
-                              part->setInstrument(*instr, 0);
-                              }
-                        else if (!ref.compare("snt-lead-ss")) {
-                              *instr = instr->fromTemplate(Ms::searchTemplate("poly-synth"));
-                              part->setInstrument(*instr, 0);
-                              }
-                        else if (!ref.compare("f-bass5")) {
-                              *instr = instr->fromTemplate(Ms::searchTemplate("bass-guitar"));
-                              part->setInstrument(*instr, 0);
-                              }
-                        else if (!ref.compare("snt-bass-ss")) {
-                              *instr = instr->fromTemplate(Ms::searchTemplate("metallic-synth"));
-                              part->setInstrument(*instr, 0);
-                              }
-                        else if (!ref.compare("mrcs")) {
-                              *instr = instr->fromTemplate(Ms::searchTemplate("maracas"));
-                              part->setInstrument(*instr, 0);
-                              }
-                        else if (!ref.compare("drmkt")) {
-                              *instr = instr->fromTemplate(Ms::searchTemplate("drumset"));
-                              part->setInstrument(*instr, 0);
-                              }
-                        else
-                              qDebug() << "WARINNG: unhandled playback instrument" << ref << "- defaulting.";
-                        }
-                  else if (nodeName == "Properties") {
-                        QDomNode currentProperty = currentNode.firstChild();
-                        while (!currentProperty.isNull()) {
-                              QString propertyName = currentProperty.attributes().namedItem("name").toAttr().value();
-                              if (!propertyName.compare("Tuning")) {
-                                    // set up the tuning for the part
-                                    QString tuningString = currentProperty.firstChild().toElement().text();
-                                    QStringList tuningStringList = tuningString.split(" ");
-                                    int strings = 0;
-                                    int tuning[tuningStringList.length()];
-                                    int frets   = 21;
-                                    for (auto iter = tuningStringList.begin(); iter != tuningStringList.end(); ++iter) {
-                                          int currentString = (*iter).toInt();
-                                          tuning[strings] = currentString;
-                                          strings++;
-                                          }
-                                          StringData* stringData = new StringData(frets, strings, tuning);
-                                          Instrument* instr = part->instr();
-                                          instr->setStringData(*stringData);
-                                    }
-                              else if (!propertyName.compare("DiagramCollection")) {
-                                    QDomNode items = currentProperty.firstChild();
-                                    QDomNode currentItem = items.firstChild();
-                                    while (!currentItem.isNull()) {
-                                          readChord(&currentItem, trackCounter);
-                                          currentItem = currentItem.nextSibling();
-                                          }
-                                    }
-                              else
-                                    unhandledNode(nodeName);
-                              currentProperty = currentProperty.nextSibling();
-                              }
-                        }
-                  else
-                        unhandledNode(nodeName);
-                  currentNode = currentNode.nextSibling();
-                  }
-
-            // add in a new part
-            part->insertStaff(s);
-            score->staves().push_back(s);
-            score->appendPart(part);
-            trackCounter++;
-            nextTrack = nextTrack.nextSibling();
-            }
-      // set the number of staves we need
-      staves = score->staves().length();
-      }
-
-//---------------------------------------------------------
-//   getNode
-//---------------------------------------------------------
-
-QDomNode GuitarPro6::getNode(QString id, QDomNode currentNode)
-      {
-      while (!(currentNode).isNull()) {
-            QString currentId = currentNode.attributes().namedItem("id").toAttr().value();
-            if (id.compare(currentId) == 0) {
-                  return currentNode;
-                  }
-            currentNode = (currentNode).nextSibling();
-            }
-      qDebug("WARNING: A null node was returned when search for an identifier. Your Guitar Pro file may be corrupted.");
-      return currentNode;
-      }
-
-//---------------------------------------------------------
-//   findNumMeasures
-//---------------------------------------------------------
-
-int GuitarPro6::findNumMeasures(GPPartInfo* partInfo)
-      {
-      QDomNode finalMasterBar = partInfo->masterBars.nextSibling();
-      while (!finalMasterBar.isNull()) {
-            GpBar gpBar;
-            gpBar.keysig = finalMasterBar.firstChild().toElement().text().toInt();
-            QString timeSignature = finalMasterBar.firstChild().nextSibling().toElement().text();
-            QList<QString> timeSignatureList = timeSignature.split("/");
-            gpBar.timesig = Fraction(timeSignatureList.first().toInt(), timeSignatureList.last().toInt());
-            bars.append(gpBar);
-            if (finalMasterBar.nextSibling().isNull())
-                  break;
-            finalMasterBar = finalMasterBar.nextSibling();
-            }
-      QString bars = finalMasterBar.lastChildElement("Bars").toElement().text();
-      //work out the number of measures (add 1 as couning from 0, and divide by number of parts)
-      int numMeasures = (bars.split(" ").last().toInt() + 1) / score->parts().length();
-      return numMeasures;
-      }
-
-//---------------------------------------------------------
-//   rhythmToDuration
-//---------------------------------------------------------
-
-Fraction GuitarPro6::rhythmToDuration(QString value)
-      {
-      Fraction l;
-      if (value.compare("Whole") == 0)
-            l.set(1, 1);
-      else if (value.compare("Half") == 0)
-            l.set(1, 2);
-      else if (value.compare("Quarter") == 0)
-            l.set(1, 4);
-      else if (value.compare("Eighth") == 0)
-            l.set(1,8);
-      else if (value.compare("16th") == 0)
-            l.set(1,16);
-      else if (value.compare("32nd") == 0)
-            l.set(1,32);
-      else if (value.compare("64th") == 0)
-            l.set(1,64);
-      else if (value.compare("128th") == 0)
-            l.set(1,128);
-      else
-            qFatal( "Error - unknown note length: %s", qPrintable(value));
-      return l;
-      }
-
-//---------------------------------------------------------
-//   readDrumNote
-//---------------------------------------------------------
-
-void GuitarPro6::readDrumNote(Note* note, int element, int variation)
-      {
-      int octaveInt;
-      int toneInt;
-      /* These numbers below were determined by creating all drum
-       * notes in a GPX format file and then analyzing the score.gpif
-       * file which specifies the score. */
-      if (element == 11 && variation == 0) {
-            octaveInt = 5;
-            toneInt = 0;
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-      }
-      else if (element == 0 && variation == 0) {
-            octaveInt = 5;
-            toneInt = 5;
-      }
-      else if (element == 5 && variation == 0) {
-            octaveInt = 5;
-            toneInt = 7;
-      }
-      else if (element == 6 && variation == 0) {
-            octaveInt = 5;
-            toneInt = 9;
-      }
-      else if (element == 7 && variation == 0) {
-            octaveInt = 5;
-            toneInt = 11;
-      }
-      else if (element == 1 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 0;
-      }
-      else if (element == 1 && variation == 1) {
-            octaveInt = 6;
-            toneInt = 0;
-            note->setHeadGroup(NoteHead::Group::HEAD_MI);
-      }
-      else if (element == 1 && variation == 2) {
-            octaveInt = 6;
-            toneInt = 0;
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-      }
-      else if (element == 8 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 2;
-      }
-      else if (element == 9 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 4;
-      }
-      else if (element == 2 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 4;
-            note->setHeadGroup(NoteHead::Group::HEAD_TRIANGLE);
-      }
-      else if (element == 15 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 5;
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-      }
-      else if (element == 15 && variation == 1) {
-            octaveInt = 6;
-            toneInt = 5;
-            note->setHeadGroup(NoteHead::Group::HEAD_DIAMOND);
-      }
-      else if (element == 15 && variation == 2) {
-            octaveInt = 6;
-            toneInt = 5;
-            note->setHeadGroup(NoteHead::Group::HEAD_MI);
-      }
-      else if (element == 3 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 5;
-            note->setHeadGroup(NoteHead::Group::HEAD_TRIANGLE);
-      }
-      else if (element == 10 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 7;
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-      }
-      else if (element == 10 && variation == 1) {
-            octaveInt = 6;
-            toneInt = 7;
-            note->setHeadGroup(NoteHead::Group::HEAD_SLASH);
-      }
-      else if (element == 10 && variation == 2) {
-            octaveInt = 6;
-            toneInt = 7;
-            note->setHeadGroup(NoteHead::Group::HEAD_XCIRCLE);
-      }
-      else if (element == 12 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 7;
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-      }
-      else if (element == 4 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 7;
-            note->setHeadGroup(NoteHead::Group::HEAD_TRIANGLE);
-      }
-      else if (element == 14 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 9;
-            note->setHeadGroup(NoteHead::Group::HEAD_CROSS);
-      }
-      else if (element == 13 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 9;
-            note->setHeadGroup(NoteHead::Group::HEAD_LA);
-      }
-      else if (element == 16 && variation == 0) {
-            octaveInt = 6;
-            toneInt = 11;
-            note->setHeadGroup(NoteHead::Group::HEAD_NORMAL);
-      }
-      // multiply octaves by 12 as 12 semitones in octave
-      note->setPitch((octaveInt * 12) + toneInt);
-      }
-
-int GuitarPro6::readBeats(QString beats, GPPartInfo* partInfo, Measure* measure, int tick, int staffIdx, int voiceNum, Tuplet* tuplets[])
-      {
-            auto currentBeatList = beats.split(" ");
-            for (auto currentBeat = currentBeatList.begin(); currentBeat != currentBeatList.end(); currentBeat++) {
-                  Fraction l;
-                  int dotted = 0;
-                  QDomNode beat = getNode(*currentBeat, partInfo->beats);
-
-                  Segment* segment = measure->getSegment(Segment::Type::ChordRest, tick);
-                  QDomNode currentNode = beat.firstChild();
-                  bool noteSpecified = false;
-                  ChordRest* cr = segment->cr(staffIdx * VOICES + voiceNum);
-                  bool tupletSet = false;
-                  Tuplet* tuplet = tuplets[staffIdx * 2 + voiceNum];
-                  while (!currentNode.isNull()) {
-                        if (currentNode.nodeName() == "Notes") {
-                              noteSpecified = true;
-                              auto notesList = currentNode.toElement().text().split(" ");
-
-                              // this could be set by rhythm if we dealt with a tuplet
-                              if (!cr)
-                                    cr = new Chord(score);
-                              cr->setTrack(staffIdx * VOICES + voiceNum);
-                              cr->setDuration(l);
-                              TDuration d(l);
-                              d.setDots(dotted);
-                              cr->setDurationType(d);
-
-                              for (auto iter = notesList.begin(); iter != notesList.end(); ++iter) {
-                                    // we have found a note
-                                    QDomNode note = getNode(*iter, partInfo->notes);
-                                    QDomNode currentNode = (note).firstChild();
-                                    // if a <Notes> tag is used but there is no <Note>, then we add a rest. This flag will allow us to check this.
-                                    while (!currentNode.isNull()) {
-                                          if (currentNode.nodeName() == "Properties") {
-                                                QDomNode currentProperty = currentNode.firstChild();
-                                                // these should not be in this scope - they may not even exist.
-                                                QString stringNum;
-                                                QString fretNum;
-                                                QString tone;
-                                                QString octave;
-                                                QString midi;
-                                                QString element;
-                                                QString variation;
-                                                while (!currentProperty.isNull()) {
-                                                      QString argument = currentProperty.attributes().namedItem("name").toAttr().value();
-                                                      if (argument == "String")
-                                                            stringNum = currentProperty.firstChild().toElement().text();
-                                                      else if (argument == "Element")
-                                                            element = currentProperty.firstChild().toElement().text();
-                                                      else if (argument == "Variation")
-                                                            variation = currentProperty.firstChild().toElement().text();
-                                                      else if (argument == "Fret")
-                                                            fretNum = currentProperty.firstChild().toElement().text();
-                                                      else if (argument == "Tone")
-                                                            tone = currentProperty.firstChild().toElement().text();
-                                                      else if (argument == "Octave")
-                                                            octave = currentProperty.firstChild().toElement().text();
-                                                      else if (argument == "Midi")
-                                                            midi = currentProperty.firstChild().toElement().text();
-                                                      else
-                                                            qDebug() << "WARNING: Not handling node argument: " << argument << "in node" << currentNode.nodeName();
-                                                      currentProperty = currentProperty.nextSibling();
-                                                }
-
-                                                if (midi != "") {
-                                                      Note* note = new Note(score);
-                                                      static_cast<Chord*>(cr)->add(note);
-                                                      note->setPitch(midi.toInt());
-                                                      note->setTpcFromPitch();
-                                                      if(!segment->cr(staffIdx * VOICES + voiceNum))
-                                                            segment->add(cr);
-                                                }
-                                                else if (element != "") {
-                                                      Note* note = new Note(score);
-                                                      static_cast<Chord*>(cr)->add(note);
-                                                      readDrumNote(note, element.toInt(), variation.toInt());
-                                                      note->setTpcFromPitch();
-                                                      if(!segment->cr(staffIdx * VOICES + voiceNum))
-                                                            segment->add(cr);
-                                                }
-                                                else if (stringNum != "") {
-                                                      Note* note = new Note(score);
-                                                      static_cast<Chord*>(cr)->add(note);
-                                                      Staff* staff = note->staff();
-
-                                                      int string =  stringNum.toInt();
-                                                      int fretNumber = fretNum.toInt();
-                                                      int musescoreString = staff->part()->instr()->stringData()->strings() - 1 - string;
-                                                      auto pitch = staff->part()->instr()->stringData()->getPitch(musescoreString, fretNumber);
-                                                      note->setFret(fretNumber);
-                                                      // we need to turn this string number for GP to the the correct string number for musescore
-                                                      note->setString(musescoreString);
-                                                      note->setPitch(pitch);
-                                                      note->setTpcFromPitch();
-                                                      if(!segment->cr(staffIdx * VOICES + voiceNum))
-                                                            segment->add(cr);
-                                                }
-                                                else if (tone != "") {
-                                                      Note* note = new Note(score);
-                                                      static_cast<Chord*>(cr)->add(note);
-                                                      int toneInt = tone.toInt();
-                                                      int octaveInt = octave.toInt();
-                                                      // multiply octaves by 12 as 12 semitones in octave
-                                                      note->setPitch((octaveInt * 12) + toneInt);
-                                                      note->setTpcFromPitch();
-                                                      if(!segment->cr(staffIdx * VOICES + voiceNum))
-                                                            segment->add(cr);
-                                                }
-
-                                                currentNode = currentNode.nextSibling();
-                                          }
-                                          else
-                                                unhandledNode(currentNode.nodeName());
-                                          currentNode = currentNode.nextSibling();
-                                    }
-                              }
-                        }
-                        else if (!currentNode.nodeName().compare("Chord")) {
-                              int key = currentNode.toElement().text().toInt();
-                              segment->add(fretDiagrams[key]);
-                              }
-                        else if (currentNode.nodeName() == "Rhythm") {
-                              // we have found a rhythm
-                              QString refString = currentNode.attributes().namedItem("ref").toAttr().value();
-                              QDomNode rhythm = getNode(refString, partInfo->rhythms);
-                              QDomNode currentNode = (rhythm).firstChild();
-                              while (!currentNode.isNull()) {
-                                    if (currentNode.nodeName() == "NoteValue") {
-                                          l = rhythmToDuration(currentNode.toElement().text());
-                                    }
-                                    else if (currentNode.nodeName() == "AugmentationDot") {
-                                          dotted = currentNode.attributes().namedItem("count").toAttr().value().toInt();
-                                          for (int count = 1; count <= dotted; count++)
-                                                l = l + (l / pow(2, count));
-                                    }
-                                    else if (currentNode.nodeName() == "PrimaryTuplet") {
-                                          qDebug("WARNING: Not handling tuplet note.");
-                                          tupletSet = true;
-                                          cr = new Chord(score);
-                                          cr->setTrack(staffIdx * VOICES + voiceNum);
-                                          if ((tuplet == 0) || (tuplet->elementsDuration() == tuplet->baseLen().fraction() * tuplet->ratio().numerator())) {
-                                                tuplet = new Tuplet(score);
-                                                tuplets[staffIdx * 2 + voiceNum] = tuplet;
-                                                tuplet->setParent(measure);
-                                          }
-                                          tuplet->setTrack(cr->track());
-                                          tuplet->setBaseLen(l);
-                                          tuplet->setRatio(Fraction(currentNode.attributes().namedItem("num").toAttr().value().toInt(),currentNode.attributes().namedItem("den").toAttr().value().toInt()));
-                                          tuplet->add(cr);
-                                    }
-                                    else
-                                          qDebug() << "WARNING: Not handling node: " << currentNode.nodeName();
-                                    currentNode = currentNode.nextSibling();
-                              }
-                        }
-                        else
-                              qDebug() << "WARNING: Not handling beat XML tag:" << currentNode.nodeName();
-                        currentNode = currentNode.nextSibling();
-                        dotted = 0;
-                  }
-                  // we have handled the note - was there a note?
-                  if (!noteSpecified) {
-                        // add a rest with length of l
-                        cr = new Rest(score);
-                        cr->setTrack(staffIdx * VOICES + voiceNum);
-                        if (tupletSet)
-                              tuplet->add(cr);
-                        TDuration d(l);
-                        cr->setDuration(l);
-                        cr->setDurationType(d);
-                        if(!segment->cr(staffIdx * VOICES + voiceNum))
-                              segment->add(cr);
-                  }
-                  tick += cr->actualTicks();
-            }
-            return tick;
-      }
-
-//---------------------------------------------------------
-//   readBars
-//---------------------------------------------------------
-
-void GuitarPro6::readBars(QDomNode* barList, Measure* measure, ClefType oldClefId[], GPPartInfo* partInfo, KeySig* t)
-      {
-      // unique bar identifiers are represented as a space separated string of numbers
-      QStringList barsString = barList->toElement().text().split(" ");
-      int staffIdx = 0;
-
-      // used to keep track of tuplets
-      Tuplet* tuplets[staves * 2];
-      for (int track = 0; track < staves*2; ++track)
-            tuplets[track] = 0;
-
-      // iterate through all the bars that have been specified
-      for (auto iter = barsString.begin(); iter != barsString.end(); ++iter) {
-            int tick = measure->tick();
-
-            t->setTrack(0);
-            Segment* s = measure->getSegment(Segment::Type::KeySig, tick);
-            s->add(t);
-
-            QDomNode barNode = getNode(*iter, partInfo->bars);
-            QDomNode currentNode = (barNode).firstChild();
-            QDomNode voice;
-            while (!currentNode.isNull()) {
-                  voice.clear();
-                  // get the clef of the bar and apply
-                  if (!currentNode.nodeName().compare("Clef")) {
-                        QString clefString = currentNode.toElement().text();
-                        ClefType clefId;
-                        if (!clefString.compare("F4"))
-                              clefId = ClefType::F8;
-                        else if (!clefString.compare("G2"))
-                              clefId = ClefType::G3;
-                        else if (!clefString.compare("Neutral"))
-                              clefId = ClefType::PERC;
-                        else
-                              qDebug() << "WARNING: unhandled clef type: " << clefString;
-                        Clef* newClef = new Clef(score);
-                        newClef->setClefType(clefId);
-                        newClef->setTrack(staffIdx * VOICES);
-                        Segment* segment = measure->getSegment(Segment::Type::Clef, 0);
-                        // only add the clef to the bar if it differs from previous measure
-                        if (measure->prevMeasure()) {
-                              if (clefId != oldClefId[staffIdx]) {
-                                    segment->add(newClef);
-                                    newClef->staff()->setClef(0, newClef->clefTypeList());
-                                    oldClefId[staffIdx] = clefId;
-                                    }
-                              }
-                        else  {
-                              segment->add(newClef);
-                              newClef->staff()->setClef(0, newClef->clefTypeList());
-                              oldClefId[staffIdx] = clefId;
-                              }
-                        }
-                  // a repeated bar (simile marking)
-                  else if (!currentNode.nodeName().compare("SimileMark")) {
-                        if (!currentNode.toElement().text().compare("Simple") ||
-                            !currentNode.toElement().text().compare("FirstOfDouble") ||
-                            !currentNode.toElement().text().compare("SecondOfDouble"))
-                              measure->cmdInsertRepeatMeasure(staffIdx);
-                        else
-                              qDebug() << "WARNING: unhandle similie mark type: " << currentNode.toElement().text();
-                        }
-                  // new voice specification
-                  else if (!currentNode.nodeName().compare("Voices")) {
-                        QString voicesString = currentNode.toElement().text();
-                        auto currentVoice = voicesString.split(" ").first();
-                        // if the voice is not -1 then we set voice
-                        if (currentVoice.compare("-1"))
-                              voice = getNode(currentVoice, partInfo->voices);
-                        }
-                  else
-                        unhandledNode(currentNode.nodeName());
-                  // go to the next node in the tree
-                  currentNode = currentNode.nextSibling();
-                  }
-
-            int voiceNum = 0;
-            if (voice.isNull()) {
-                  qDebug("NULL voice detected!");
-                  Fraction l = Fraction(1,1);
-                  // add a rest with length of l
-                  ChordRest* cr = new Rest(score);
-                  cr->setTrack(staffIdx * VOICES + voiceNum);
-                  TDuration d(l);
-                  cr->setDuration(l);
-                  cr->setDurationType(d);
-                  Segment* segment = measure->getSegment(Segment::Type::ChordRest, tick);
-                  if(!segment->cr(staffIdx * VOICES + voiceNum))
-                        segment->add(cr);
-                  tick += cr->actualTicks();
-                  staffIdx++;
-                  continue;
-                  }
-
-            // read the beats that occur in the bar
-            tick = readBeats(voice.firstChild().toElement().text(), partInfo, measure, tick, staffIdx, voiceNum, tuplets);
-            // increment the counter for parts
-            staffIdx++;
-            }
-      }
-
-//---------------------------------------------------------
-//   readMasterBars
-//---------------------------------------------------------
-
-void GuitarPro6::readMasterBars(GPPartInfo* partInfo)
-      {
-      Measure* measure = score->firstMeasure();
-      int bar = 0;
-      QDomNode nextMasterBar = partInfo->masterBars;
-      nextMasterBar = nextMasterBar.nextSibling();
-      int measureCounter = 0;
-      ClefType oldClefId[staves];
-      do    {
-            const GpBar& gpbar = bars[bar];
-
-            if (!gpbar.marker.isEmpty()) {
-                  Text* s = new RehearsalMark(score);
-                  s->setText(gpbar.marker.trimmed());
-                  s->setTrack(0);
-                  Segment* segment = measure->getSegment(Segment::Type::ChordRest, measure->tick());
-                  segment->add(s);
-                  }
-
-            QDomNode key = nextMasterBar.firstChild();
-            KeySig* t = new KeySig(score);
-            t->setKey(Key(key.firstChild().toElement().text().toInt()));
-
-            QDomNode time = key.nextSibling();
-            measure->setTimesig(bars[measureCounter].timesig);
-            measure->setLen(bars[measureCounter].timesig);
-            QDomNode barList = time.nextSibling();
-            readBars(&barList, measure, oldClefId, partInfo, t);
-            measureCounter++;
-            nextMasterBar = nextMasterBar.nextSibling();
-            measure = measure->nextMeasure();
-            bar++;
-            } while (!nextMasterBar.isNull());
-      }
-
-//---------------------------------------------------------
-//   readGpif
-//---------------------------------------------------------
-
-void GuitarPro6::readGpif(QByteArray* data)
-      {
-      QDomDocument qdomDoc;
-      qdomDoc.setContent(*data);
-      QDomElement qdomElem = qdomDoc.documentElement();
-      // GPRevision node
-      QDomNode revision = qdomElem.firstChild();
-      // Score node
-      QDomNode scoreNode = revision.nextSibling();
-      readScore(&scoreNode);
-      // MasterTrack node
-      QDomNode masterTrack = scoreNode.nextSibling();
-      readMasterTracks(&masterTrack);
-      // Tracks node
-      QDomNode eachTrack = masterTrack.nextSibling();
-      readTracks(&eachTrack);
-      // MasterBars node
-      GPPartInfo partInfo;
-      QDomNode masterBars = eachTrack.nextSibling();
-      QDomNode bars = masterBars.nextSibling();
-      QDomNode voices = bars.nextSibling();
-      QDomNode beats = voices.nextSibling();
-      QDomNode notes = beats.nextSibling();
-      QDomNode rhythms = notes.nextSibling();
-
-      // set up the partInfo struct to contain information from the file
-      partInfo.masterBars = masterBars.firstChild();
-      partInfo.bars = bars.firstChild();
-      partInfo.voices = voices.firstChild();
-      partInfo.beats = beats.firstChild();
-      partInfo.notes = notes.firstChild();
-      partInfo.rhythms = rhythms.firstChild();
-
-      measures = findNumMeasures(&partInfo);
-
-      createMeasures();
-      readMasterBars(&partInfo);
-      // set the starting tempo of the score
-      setTempo(/*tempo*/120, score->firstMeasure());
-      }
-
-//---------------------------------------------------------
-//   read
-//---------------------------------------------------------
-
-void GuitarPro6::read(QFile* fp)
-      {
-      qDebug("reading guitar pro v6 file (.gpx)...");
-      f = fp;
-      this->buffer = new QByteArray();
-      *(this->buffer) = fp->readAll();
-      // decompress and read files contained within GPX file
-      readGPX(this->buffer);
-      delete this->buffer;
-      }
-
-//---------------------------------------------------------
-//   readBeatEffects
-//---------------------------------------------------------
-
-int GuitarPro6::readBeatEffects(int, Segment*)
-      {
-      qDebug("reading beat effects (.gpx)...");
-      return 0;
-      }
-
 
 //---------------------------------------------------------
 //   importGTP
